requirements.txt
flask
pillow
fpdf2
sortablejs